+++ packages/server/src/controllers/credentials/index.ts
@@ -28,8 +28,14 @@ const deleteCredentials = async (req: Request, res: Response, next: NextFunction
                 `Error: credentialsController.deleteCredentials - id not provided!`
             )
         }
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await credentialsService.deleteCredentials(req.params.id, workspaceId)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: credentialsController.deleteCredentials - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await credentialsService.deleteCredentials(req.params.id, workspaceId)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
@@ -38,8 +44,14 @@ const apiResponse = await credentialsService.deleteCredentials(req.params.id, wo
 
 const getAllCredentials = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await credentialsService.getAllCredentials(req.query.credentialName, workspaceId)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: credentialsController.getAllCredentials - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await credentialsService.getAllCredentials(req.query.credentialName, workspaceId)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
@@ -54,8 +66,14 @@ const getCredentialById = async (req: Request, res: Response, next: NextFunction
                 `Error: credentialsController.getCredentialById - id not provided!`
             )
         }
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await credentialsService.getCredentialById(req.params.id, workspaceId)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: credentialsController.getCredentialById - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await credentialsService.getCredentialById(req.params.id, workspaceId)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
@@ -76,8 +94,14 @@ const updateCredential = async (req: Request, res: Response, next: NextFunction)
                 `Error: credentialsController.updateCredential - body not provided!`
             )
         }
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await credentialsService.updateCredential(req.params.id, req.body, workspaceId)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: credentialsController.updateCredential - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await credentialsService.updateCredential(req.params.id, req.body, workspaceId)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
